LayoutTests/fast/scrolling/programmatic-smooth-then-immediate-cancels-animation.html

<!DOCTYPE html> <!-- webkit-test-runner [ AsyncOverflowScrollingEnabled=true ] -->
<html>
<head>
    <style>
        .scroller {
            height: 400px;
            width: 400px;
            border: 1px solid black;
            overflow-y: scroll;
        }

        .contents {
            width: 100%;
            height: 400%;
            background-image: repeating-linear-gradient(to bottom, white, silver 250px);
        }
    </style>
    <script src="../../resources/ui-helper.js"></script>
    <script src="../../resources/js-test.js"></script>
    <script>
        jsTestIsAsync = true;

        let scrollEventCount = 0;

        addEventListener("load", async () => {
            await UIHelper.delayFor(0);

            scrollContainer = document.querySelector(".scroller");
            scrollContainer.scrollTop = 200;
            scrollContainer.scrollTo({ top: 500, behavior: 'smooth' });

            await UIHelper.renderingUpdate();

            scrollContainer.addEventListener('scroll', () => {
                ++scrollEventCount;
            }, false);

            scrollContainer.scrollTo({ top: 100 });
            await UIHelper.renderingUpdate();
            shouldBe('scrollContainer.scrollTop', '100');
            
            // Check that the scroll didn't continue to animate.
            await UIHelper.renderingUpdate();
            shouldBe('scrollContainer.scrollTop', '100');
            
            await UIHelper.waitForTargetScrollAnimationToSettle(scrollContainer);

            shouldBeTrue('scrollEventCount <= 2');

            finishJSTest();
        });
    </script>
</head>
<body>
    <div class="scroller">
        <div class="contents"></div>
    </div>
    <div id="console"></div>
</body>
</html>
