LayoutTests/fast/scrolling/programmatic-smooth-then-immediate-cancels-animation-expected.txt

PASS scrollContainer.scrollTop is 100
PASS scrollContainer.scrollTop is 100
PASS scrollEventCount <= 2 is true
PASS successfullyParsed is true

TEST COMPLETE
